basic home page
little blurb of text about me and a nice picture of me. Tailwind's basis feature seems quite helpful as of now, and I had struggled a lot with resizing the image. Added some custom scaling in tailwind.config.js but ultimately threw it away because it was messing up the screen ratio: the image was scalled correctly but the screen had lots of empty space. Thus, I went with h-sreen and h-full