basic navbar
adding on the navbar ruins h-screen of the content because of the additional padding. I can't seem to find a good alternative online and ended up going with https://stackoverflow.com/questions/72673561/tailwind-h-screen-container-is-header-amount-of-height-too-far-off-the-bottom to "hard-code" the problem. only problem is that my content is slightly off center now... One way I can fix this is by incorporating the navbar dierctly into each page rather than the layout. It's bad code because of repetitiveness but I don't plan on having too many pages anyways. I despise navbars...