Projects Page
I wasn't sure if I wanted to make this, but I'm happy with how it turned out. There's so many little side projects that I've done but kind of forgot about. This is a nice place to put them. Ideally adjust the bottom spacing for each ProjectCard, but basically everything is really well done in terms of programming!. Features include description on image hover, hyperlink on image title, and little buttons for relevant skills